+++ tasks/lib/commands.js
@@ -1,5 +1,6 @@
 'use strict';
 
+var _ = require('lodash');
 var util = require('util');
 var conf = require('./conf.js');
 var parse = require('./parse.js');
@@ -10,14 +11,17 @@ module.exports = {
         return util.format('%s && sudo pm2 reload all || echo "pm2 not started."', running);
     },
     pm2_start: function (name) {
-        var args = parse.args({
+        var defaults = {
             NODE_ENV: name
-        },'--');
-        var app_env = parse.args(
-            conf('ENV')
-        );
-        return util.format('%s || sudo %s pm2 start %s/%s -i 2 --name %s -- %s || echo "pm2 already started."',
-            running, app_env, conf('SRV_CURRENT'), conf('NODE_SCRIPT'), name, args
+        };
+        var user = conf('ENV');
+        var env = {};
+
+        // user can override NODE_ENV if need be
+        _.assign(env, defaults, user);
+
+        return util.format('%s || sudo %s pm2 start %s/%s -i 2 --name %s || echo "pm2 already started."',
+            running, env, conf('SRV_CURRENT'), conf('NODE_SCRIPT'), name
         );
     }
 };